+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixed URI parsing for IPv6 literals containing embedded IPv4 addresses
 - Fixed malformed UTF-8 URI strings being parsed as empty URIs
 
 ## 2.10.2 - 2026-05-25